+++ hey-i-am-still-working/cmake/compiler_msvc.cmake
@@ -1,13 +1,21 @@
 
-option(MSVC_ENABLE_PARALLEL_BUILD "If enabled, build multiple files in parallel." ON)
+option(HEY_I_AM_STILL_WORKING_USE_MSVC_PARALLEL_BUILD "If enabled, build multiple files in parallel." ON)
+option(HEY_I_AM_STILL_WORKING_USE_MSVC_STATIC_ANALYSIS "If enabled, run MSVC built-in static analysis and generate appropriate warnings" OFF)
+option(HEY_I_AM_STILL_WORKING_USE_WIN32_LEAN_AND_MEAN "If enabled, define WIN32_LEAN_AND_MEAN" ON)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /permissive-")
+if(HEY_I_AM_STILL_WORKING_NOT_SUBPROJECT)
+  message(STATUS "hashids compiler MSVC global conf is in active")
 
-# Force generating debugging symbols in Release build.
-set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /Zi")
-set(CMAKE_EXE_LINKER_FLAGS_RELEASE "${CMAKE_EXE_LINKER_FLAGS_RELEASE} /DEBUG /OPT:REF /OPT:ICF")
+  # Force generating debugging symbols in Release build.
+  set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /Zi")
+  set(CMAKE_EXE_LINKER_FLAGS_RELEASE "${CMAKE_EXE_LINKER_FLAGS_RELEASE} /DEBUG /OPT:REF /OPT:ICF")
+endif()
 
-function(apply_hey_i_am_still_working_compile_conf TARGET)
+message(STATUS "HEY_I_AM_STILL_WORKING_USE_MSVC_PARALLEL_BUILD = ${HEY_I_AM_STILL_WORKING_USE_MSVC_PARALLEL_BUILD}")
+message(STATUS "HEY_I_AM_STILL_WORKING_USE_MSVC_STATIC_ANALYSIS = ${HEY_I_AM_STILL_WORKING_USE_MSVC_STATIC_ANALYSIS}")
+message(STATUS "HEY_I_AM_STILL_WORKING_USE_WIN32_LEAN_AND_MEAN = ${HEY_I_AM_STILL_WORKING_USE_WIN32_LEAN_AND_MEAN}")
+
+function(hey_i_am_still_working_apply_common_compile_options TARGET)
   target_compile_definitions(${TARGET}
     PUBLIC
       _UNICODE
@@ -17,6 +25,8 @@ function(apply_hey_i_am_still_working_compile_conf TARGET)
       $<$<CONFIG:DEBUG>:
         _DEBUG
       >
+
+      $<$<BOOL:HEY_I_AM_STILL_WORKING_USE_WIN32_LEAN_AND_MEAN>:WIN32_LEAN_AND_MEAN>
   )
 
   target_compile_options(${TARGET}
@@ -24,34 +34,49 @@ function(apply_hey_i_am_still_working_compile_conf TARGET)
       /W4
       /wd4819 # source characters not in current code page.
 
-      /Zc:inline # Have the compiler eliminate unreferenced COMDAT functions and data before emitting the object file.
+      /Zc:inline            # Have the compiler eliminate unreferenced COMDAT functions and data before emitting the object file.
+      /Zc:rvalueCast        # Enforce the standard rules for explicit type conversion.
+      /Zc:strictStrings     # Don't allow conversion from a string literal to mutable characters.
+      /Zc:threadSafeInit    # Enable thread-safe function-local statics initialization.
 
-      $<$<BOOL:${MSVC_ENABLE_PARALLEL_BUILD}>:/MP>
-
-    PUBLIC
-      /Zc:referenceBinding # Disallow temporaries from binding to non-const lvalue references.
-      /Zc:rvalueCast # Enforce the standard rules for explicit type conversion.
-      /Zc:implicitNoexcept # Enable implicit noexcept specifications where required, such as destructors.
-      /Zc:strictStrings # Don't allow conversion from a string literal to mutable characters.
-      /Zc:threadSafeInit # Enable thread-safe function-local statics initialization.
-      /Zc:throwingNew # Assume operator new throws on failure.
-
-      /permissive- # Be mean, don't allow bad non-standard stuff (C++/CLI, __declspec, etc. are all left intact).
+      /permissive-  # Be mean, don't allow bad non-standard stuff (C++/CLI, __declspec, etc. are all left intact).
   )
 endfunction()
 
-# Enable static analysis for all targets could result in excessive warnings.
-# Thus we decided to enable it for targets only we explicitly specify.
-# Use /wd args to suppress analysis warnings we cannot resolve.
-function(enable_hey_i_am_still_working_msvc_static_analysis_conf TARGET)
+function(hey_i_am_still_working_apply_msvc_parallel_build TARGET)
+  message(STATUS "Apply hashids msvc parallel build for ${TARGET}")
+
+  target_compile_options(${TARGET}
+    PRIVATE
+      /MP
+  )
+endfunction()
+
+# To explicitly suppress spcific warnings:
+# hey_i_am_still_working_apply_msvc_static_analysis(foobar
+#   WDL
+#     /wd6011
+# )
+function(hey_i_am_still_working_apply_msvc_static_analysis TARGET)
+  message(STATUS "Apply hashids msvc static analysis for ${TARGET}")
+
   set(multiValueArgs WDL)
   cmake_parse_arguments(ARG "" "" "${multiValueArgs}" ${ARGN})
 
   target_compile_options(${TARGET}
     PRIVATE
       /analyze
-      /analyze:WX-
+
+      /wd6001 # Using uninitialized memory.
+      /wd6011 # Dereferencing potentially NULL pointer.
 
       ${ARG_WDL}
   )
+
+  if (CMAKE_GENERATOR MATCHES "Visual Studio")
+    set_target_properties(${TARGET} PROPERTIES
+      VS_GLOBAL_EnableMicrosoftCodeAnalysis true
+      VS_GLOBAL_RunCodeAnalysis true
+    )
+  endif()
 endfunction()
